+++ webim/src/Router.php
@@ -277,6 +277,7 @@ EOF;
                 'success' => true,
                 'connection' => $data->connection,
                 'user' => $this->user,
+                'presences' => $data->presences,
                 'buddies' => array_values($rtBuddies),
                 'rooms' => array_values($rtRooms),
                 'new_messages' => $offlineMessages,
@@ -312,7 +313,32 @@ EOF;
 	public function buddies() {
         $uid = $this->user->id;
 		$ids = $this->input('ids');
-		$this->jsonReply($this->plugin->buddiesByIds($uid, $ids));
+        $vids = array();
+        $uids = array();
+        foreach(explode(',', $ids) as $id) {
+            if($this->isvid($id)) { 
+                $vids[] = $id;
+            } else {
+                $uids[] = $id;
+            }
+        }
+        $buddies = array_merge(
+            $this->plugin->buddiesByIds($uid, $uids),
+            $this->model->visitors($vids)
+        );
+        $buddyIds = array_map(array($this, 'buddyId'), $buddies);
+        $presences = $this->client->presences($buddyIds);
+        foreach($buddies as $buddy) {
+            $id = $buddy->id;
+            if( isset($presences->$id) ) {
+                $buddy->presence = 'online';
+                $buddy->show = $presences->$id;
+            } else {
+                $buddy->presence = 'offline';
+                $buddy->show = 'unavailable';
+            }
+        }
+		$this->jsonReply($buddies);
 	}
 
     /**
@@ -470,7 +496,7 @@ EOF;
             'id' => $room->name,
             'nick' => $room->nick,
             'temporary' => true,
-            'pic_url' => WEBIM_IMAGE('room.png')
+            'avatar' => WEBIM_IMAGE('room.png')
         ));
     }
 
@@ -495,7 +521,7 @@ EOF;
             'id' => $roomId,
             'nick' => $nick,
             'temporary' => true,
-            'pic_url' => WEBIM_IMAGE('room.png')
+            'avatar' => WEBIM_IMAGE('room.png')
         ));
 	}