+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 ==================
 
+5.5 (2014-06-17)
+------------------
+
+* Router.php: return presences to browser used by 'chatbtn'
+* upgrade static file to 5.5 and support detach window
+
 5.4.2 (2014-04-21)
 ------------------